JH-165 Refactor: move test of test_content_list_detail to each model test
