+++ v2/src/tests/test_topics.py
@@ -8,6 +8,76 @@ from django.urls import reverse
 from rest_framework import status
 
 
+class TestTopicList(TestCase):
+    def setUp(self):
+        self.c = Client()
+        self.user = CtrlfUser.objects.create_user(email="test@test.com", password="12345")
+        self.note = Note.objects.create(title="test note")
+        self.note.owners.add(self.user)
+
+    def _add_topics_to_note(self, note, count):
+        topic_list = []
+        for i in range(count):
+            topic_data = {"note": note, "title": f"test topic{i + 1} to {note.title}"}
+            topic = Topic.objects.create(**topic_data)
+            topic.owners.add(self.user)
+            topic_list.append(topic)
+        return topic_list
+
+    def _call_api(self, note_id):
+        return self.c.get(reverse("notes:topic_list", kwargs={"note_id": note_id}))
+
+    def test_topic_list_should_return_200(self):
+        # Given: 이미 저장된 topic들, 유효한 note id
+        note_id = self.note.id
+        topic_list = self._add_topics_to_note(note=self.note, count=10)
+        # When : API 실행
+        response = self._call_api(note_id)
+        # Then : 상태코드 200
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        # And  : 이미 저장된 topic 개수와 같아야 함.
+        response = response.data
+        self.assertEqual(len(response), len(topic_list))
+
+    def test_topic_list_should_return_topic_list_only_dependent_on_note_id(self):
+        # Given: note A, B를 생성한다.
+        note_A = Note.objects.create(title="note A")
+        note_B = Note.objects.create(title="note B")
+        # And: note A에 3개, note B에 5개의 topic을 생성한다.
+        self._add_topics_to_note(note=note_A, count=3)
+        self._add_topics_to_note(note=note_B, count=5)
+
+        # When: note_A의 id로 topic list API 호출한다.
+        response = self._call_api(note_A.id)
+
+        # Then: status code는 200이다,
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        # And: topic 개수는 3개이다.
+        self.assertEqual(len(response.data), 3)
+
+    def test_topic_list_should_return_200_by_empty_topic_list(self):
+        # Given: 유효한 note id
+        note_id = self.note.id
+        # When : API 실행
+        response = self._call_api(note_id)
+        # Then : 상태코드 200
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        # And  : 빈 배열을 return 해야함.
+        self.assertEqual(response.data, [])
+
+    def test_topic_list_should_return_404_by_invalid_note_id(self):
+        # Given: 이미 저장된 topic들, 유효하지 않은 note id
+        invalid_not_id = 9999
+        self._add_topics_to_note(note=self.note, count=10)
+        # When : API 실행
+        response = self._call_api(invalid_not_id)
+        # Then : 상태코드 404
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+        # And  : 메세지는 "노트를 찾을 수 없습니다." 이어야 함.
+        response = response.data
+        self.assertEqual(response["message"], "노트를 찾을 수 없습니다.")
+
+
 class TestTopicCreate(TestCase):
     def setUp(self) -> None:
         self.client = Client()
@@ -90,6 +160,39 @@ class TestTopicCreate(TestCase):
         self.assertEqual(Topic.objects.count(), 0)
         # And: Issue도 생성되지 않는다.
         self.assertEqual(Issue.objects.count(), 0)
+
+
+class TestTopicDetail(TestCase):
+    def setUp(self):
+        self.c = Client()
+        self.user = CtrlfUser.objects.create_user(email="test@test.com", password="12345")
+        self.note = Note.objects.create(title="test note")
+        self.note.owners.add(self.user)
+        self.topic = Topic.objects.create(note=self.note, title="test topic")
+        self.topic.owners.add(self.user)
+
+    def test_topic_detail_should_return_200(self):
+        # Given : 유효한 topic id, 이미 저장된 topic
+        topic_id = self.topic.id
+        # When  : API 실행
+        response = self.c.get(reverse("topics:topic_detail_update", kwargs={"topic_id": topic_id}))
+        # Then  : 상태코드 200
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        # And   : 불러온 정보가 저장된 정보와 일치해야 한다.
+        response = response.data
+        self.assertEqual(response["title"], "test topic")
+        self.assertEqual(response["note"], self.note.id)
+
+    def test_topic_detail_should_return_404_by_invalid_topic_id(self):
+        # Given : 유효하지 않은 topic id, 이미 저장된 topic
+        invalid_topic_id = 1234
+        # When  : API 실행
+        response = self.c.get(reverse("topics:topic_detail_update", kwargs={"topic_id": invalid_topic_id}))
+        # Then  : 상태코드 404
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+        # And   : 메세지는 "토픽을 찾을 수 없습니다." 이어야 한다.
+        response = response.data
+        self.assertEqual(response["message"], "토픽을 찾을 수 없습니다.")
 
 
 class TestTopicUpdate(TestCase):
